requirements.txt
asgiref==3.4.1
Django==4.0.1
ImageHash==4.2.1
mysqlclient==2.1.0
numpy==1.22.0
Pillow==9.0.0
psycopg2==2.9.3
PyWavelets==1.2.0
scipy==1.7.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
